sched: Add generic scheduler code and basic roundrobin scheduler.
* Added sys/timer.h for timed events
* Refactored timers to be arch-independent
* Added sched/scheduler.h for scheduler callbacks
* Added concrete scheduler implementation (round robin style)

+++ arch/i386/machine/regs.c
@@ -108,7 +108,7 @@ backtrace(unsigned int max_frames)
 }
 
 void
-set_stack(struct regs *regs, reg_t stack, size_t stack_size)
+set_stack(struct regs *regs, reg_t stack, unsigned long stack_size)
 {
         regs->esp = stack + stack_size - 4;
         regs->ebp = 0;
