+++ arch/i686/machine/regs.c
@@ -122,7 +122,7 @@ backtrace(unsigned int max_frames)
 }
 
 void
-set_stack(struct regs *regs, reg_t stack, size_t stack_size)
+set_stack(struct regs *regs, reg_t stack, unsigned long stack_size)
 {
         regs->rsp = stack + stack_size - 16;
         regs->rbp = 0;
